requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
coverage==4.0.2
dj-database-url==0.3.0
Django==1.8
django-bootstrap-form==3.2
django-envie==0.0.9
django-filter==0.11.0
django-postgrespool==0.3.0
funcsigs==0.4
gunicorn==19.3.0
httplib2==0.9.2
kombu==3.0.29
Markdown==2.6.4
mock==1.3.0
oauth2==1.9.0.post1
oauthlib==1.0.3
pbr==1.8.1
Pillow==3.0.0
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.13
pytz==2015.7
PyYAML==3.11
requests==2.8.1
requests-oauthlib==0.5.0
selenium==2.48.0
six==1.10.0
SQLAlchemy==1.0.9
wheel==0.24.0
whitenoise==2.0.4
